examples/base-trading-bot/requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
web3>=6.13.0